+++ packages/cortex-react/src/components/Uploader/Dropzone.tsx
@@ -19,26 +19,26 @@ export const Dropzone = ({
     <div
       {...getRootProps()}
       className={clsx(
-        'bg-white w-full border-dashed border-2 p-deca flex flex-col justify-center rounded-mili overflow-auto h-[18rem]',
+        'bg-surface-overlay w-full border-dashed border-2 p-deca flex flex-col justify-center rounded-mili overflow-auto h-[18rem]',
         {
           'border-success-medium bg-gray-100': isDragActive,
         }
       )}
     >
       <input {...getInputProps()} />
-      <div className="flex flex-col justify-center text-center items-center">
-        <div>
+      <div className="flex flex-col justify-center text-center items-center w-full">
+        <div className="flex flex-col items-center w-full">
           <HiOutlineCloudArrowUp
-            className="text-primary-medium mb-nano"
+            className="text-primary-medium mb-nano mx-auto"
             size={35}
           />
-          <p className="text-lg font-semibold" data-testid="select-dropzone">
+          <p className="text-lg font-semibold w-full text-center" data-testid="select-dropzone">
             {selectFileText}
           </p>
         </div>
         {hasButton ? (
           <div className="flex flex-col items-center gap-centi">
-            <p className="text-sm text-secondary-medium">{dropText}</p>
+            <p className="text-sm text-content-low">{dropText}</p>
             {isFileLimitReached ? (
               <Tooltip text="Você só pode selecionar um único arquivo.">
                 <button