+++ packages/react-core/src/components/molecules/PhoneInput/index.ts
@@ -1,0 +1,10 @@
+import './polyfills';
+
+export { default as PhoneInput } from './PhoneInput';
+export type { PhoneInputProps } from './PhoneInput';
+export { default as PhoneCountrySelector } from './PhoneCountrySelector';
+export type { PhoneCountrySelectorProps } from './PhoneCountrySelector';
+export { default as CountryOption } from './CountryOption';
+export { default as FlagIcon } from './FlagIcon';
+export type { FlagIconProps } from './FlagIcon';
+export * from './types';