+++ packages/react-native-kit/src/components/molecules/PhoneInput/index.ts
@@ -1,0 +1,4 @@
+export { PhoneInput } from './PhoneInput';
+export type { NativePhoneInputProps } from './PhoneInput';
+export { FlagIcon } from './FlagIcon';
+export type { NativeFlagIconProps } from './FlagIcon';